Split Time component from Timestamp to allow just rendering of the time element
